Split queue history checker models
