+++ examples/evaluators/queue/checker_test.go
@@ -39,153 +39,6 @@ func TestAccuracyTimesOutStuckCandidateOperation(t *testing.T) {
 	}
 }
 
-func enqueueRecord(client int, value uint64, ok bool, call int64) recordedOperation {
-	return recordedOperation{
-		ClientID: client,
-		Input:    queueInput{Kind: "enqueue", Value: &value},
-		Call:     call,
-		Output:   queueOutput{EnqueueOK: &ok},
-		Return:   call + 1,
-	}
-}
-
-func dequeueRecord(client int, value *uint64, call int64) recordedOperation {
-	output := queueOutput{DequeueNone: value == nil}
-	if value != nil {
-		copy := *value
-		output.DequeueVal = &copy
-	}
-	return recordedOperation{
-		ClientID: client,
-		Input:    queueInput{Kind: "dequeue"},
-		Call:     call,
-		Output:   output,
-		Return:   call + 1,
-	}
-}
-
-func value(value uint64) *uint64 {
-	return &value
-}
-
-func TestQueueModelAcceptsBoundedFIFOHistory(t *testing.T) {
-	history := []recordedOperation{
-		dequeueRecord(0, nil, 1),
-		enqueueRecord(0, 10, true, 3),
-		enqueueRecord(0, 20, true, 5),
-		enqueueRecord(0, 30, false, 7),
-		dequeueRecord(0, value(10), 9),
-		dequeueRecord(0, value(20), 11),
-		dequeueRecord(0, nil, 13),
-	}
-	if !checkExactFIFOHistory(2, history) {
-		t.Fatal("valid bounded FIFO history was rejected")
-	}
-}
-
-func TestQueueModelRejectsIncorrectHistories(t *testing.T) {
-	tests := map[string][]recordedOperation{
-		"fifo violation": {
-			enqueueRecord(0, 10, true, 1),
-			enqueueRecord(0, 20, true, 3),
-			dequeueRecord(0, value(20), 5),
-		},
-		"fabricated value": {
-			dequeueRecord(0, value(99), 1),
-		},
-		"duplicate value": {
-			enqueueRecord(0, 10, true, 1),
-			dequeueRecord(0, value(10), 3),
-			dequeueRecord(0, value(10), 5),
-		},
-		"capacity overflow": {
-			enqueueRecord(0, 10, true, 1),
-			enqueueRecord(0, 20, true, 3),
-		},
-		"early full response": {
-			enqueueRecord(0, 10, false, 1),
-		},
-	}
-
-	for name, history := range tests {
-		t.Run(name, func(t *testing.T) {
-			if checkExactFIFOHistory(1, history) {
-				t.Fatal("invalid history was accepted")
-			}
-		})
-	}
-}
-
-func TestScenarioModelsApplyDistinctReservationSemantics(t *testing.T) {
-	first := enqueueRecord(0, 10, true, 1)
-	first.Return = 8
-	full := enqueueRecord(1, 20, false, 3)
-	empty := dequeueRecord(2, nil, 5)
-	history := []recordedOperation{first, full, empty}
-
-	if checkScenarioHistory(scenarioSPSC, 1, history) {
-		t.Fatal("exact SPSC model accepted FULL and EMPTY around an unpublished enqueue")
-	}
-	if !checkScenarioHistory(scenarioMPMC, 1, history) {
-		t.Fatal("reservation-aware MPMC model rejected a reserved but unpublished enqueue")
-	}
-}
-
-func TestReservationAwareModelRejectsInvalidHistories(t *testing.T) {
-	tests := map[string][]recordedOperation{
-		"full without used capacity": {
-			enqueueRecord(0, 10, false, 1),
-		},
-		"empty with published item": {
-			enqueueRecord(0, 10, true, 1),
-			dequeueRecord(1, nil, 3),
-		},
-		"fifo violation": {
-			enqueueRecord(0, 10, true, 1),
-			enqueueRecord(1, 20, true, 3),
-			dequeueRecord(2, value(20), 5),
-		},
-		"fabricated value": {
-			dequeueRecord(0, value(99), 1),
-		},
-		"duplicate value": {
-			enqueueRecord(0, 10, true, 1),
-			dequeueRecord(1, value(10), 3),
-			dequeueRecord(2, value(10), 5),
-		},
-	}
-
-	for name, history := range tests {
-		t.Run(name, func(t *testing.T) {
-			if checkReservationAwareFIFOHistory(1, history) {
-				t.Fatal("invalid reservation-aware history was accepted")
-			}
-		})
-	}
-}
-
-func TestReservationAwareFullCountsAllUsedCapacity(t *testing.T) {
-	first := enqueueRecord(0, 10, true, 1)
-	first.Return = 8
-	full := enqueueRecord(1, 20, false, 3)
-	history := []recordedOperation{first, full}
-
-	if checkReservationAwareFIFOHistory(2, history) {
-		t.Fatal("FULL was accepted with one reservation and two slots of capacity")
-	}
-
-	published := enqueueRecord(0, 5, true, 1)
-	reserved := enqueueRecord(1, 10, true, 3)
-	reserved.Return = 10
-	full = enqueueRecord(2, 20, false, 5)
-	if !checkReservationAwareFIFOHistory(
-		2,
-		[]recordedOperation{published, reserved, full},
-	) {
-		t.Fatal("FULL was rejected when published plus reserved items reached capacity")
-	}
-}
-
 func TestWorkerCountsEnforceScenarioShape(t *testing.T) {
 	tests := []struct {
 		scenario               scenario
