+++ examples/evaluators/queue/reservation_fifo_model.go
@@ -1,75 +1,10 @@
 package main
 
-import "github.com/anishathalye/porcupine"
+import (
+	"slices"
 
-func checkScenarioHistory(s scenario, capacity int, history []recordedOperation) bool {
-	switch s {
-	case scenarioSPSC, scenarioMPSC:
-		return checkExactFIFOHistory(capacity, history)
-	case scenarioMPMC:
-		return checkReservationAwareFIFOHistory(capacity, history)
-	default:
-		return false
-	}
-}
-
-func correctnessContract(s scenario) string {
-	if s == scenarioMPMC {
-		return "reservation-aware bounded FIFO"
-	}
-	return "linearizable bounded FIFO"
-}
-
-func checkExactFIFOHistory(capacity int, history []recordedOperation) bool {
-	operations := make([]porcupine.Operation, 0, len(history))
-	for _, op := range history {
-		operations = append(operations, op.porcupine())
-	}
-	return porcupine.CheckOperations(exactFIFOModel(capacity), operations)
-}
-
-func exactFIFOModel(capacity int) porcupine.Model {
-	return porcupine.Model{
-		Init: func() any { return []uint64{} },
-		Step: func(state, input, output any) (bool, any) {
-			current := state.([]uint64)
-			in := input.(queueInput)
-			out := output.(queueOutput)
-			switch in.Kind {
-			case "enqueue":
-				if in.Value == nil || out.EnqueueOK == nil {
-					return false, state
-				}
-				expected := len(current) < capacity
-				if *out.EnqueueOK != expected {
-					return false, state
-				}
-				if !expected {
-					return true, state
-				}
-				next := append([]uint64(nil), current...)
-				next = append(next, *in.Value)
-				return true, next
-			case "dequeue":
-				if len(current) == 0 {
-					return out.DequeueNone && out.DequeueVal == nil, state
-				}
-				if out.DequeueNone || out.DequeueVal == nil || *out.DequeueVal != current[0] {
-					return false, state
-				}
-				next := append([]uint64(nil), current[1:]...)
-				return true, next
-			default:
-				return false, state
-			}
-		},
-		Equal: func(first, second any) bool {
-			a := first.([]uint64)
-			b := second.([]uint64)
-			return equalValues(a, b)
-		},
-	}
-}
+	"github.com/anishathalye/porcupine"
+)
 
 type reservationEventKind uint8
 
@@ -195,7 +130,7 @@ func reservationAwareFIFOModel(capacity int) porcupine.Model {
 		Equal: func(first, second any) bool {
 			a := first.(reservationState)
 			b := second.(reservationState)
-			if !equalValues(a.Published, b.Published) || len(a.Reserved) != len(b.Reserved) {
+			if !slices.Equal(a.Published, b.Published) || len(a.Reserved) != len(b.Reserved) {
 				return false
 			}
 			for value := range a.Reserved {
@@ -217,16 +152,4 @@ func cloneReservationState(state reservationState) reservationState {
 		Reserved:  reserved,
 		Published: append([]uint64(nil), state.Published...),
 	}
-}
-
-func equalValues(first, second []uint64) bool {
-	if len(first) != len(second) {
-		return false
-	}
-	for index := range first {
-		if first[index] != second[index] {
-			return false
-		}
-	}
-	return true
 }
